+++ App.js
@@ -1,0 +1,10 @@
+import { View, Text } from 'react-native'
+import React from 'react'
+import Navigator from './navigation/Navigator'
+
+
+export default function App() {
+    return (
+        <Navigator />
+    )
+}
